Rename long to integer

+++ rust/common/token.rs
@@ -173,7 +173,7 @@ string_enum! { ValueType
     Decimal = "decimal",
     Double = "double",
     Duration = "duration",
-    Long = "long",
+    Integer = "integer",
     String = "string",
 }
